partial implementation of native protocol
git-svn-id: file:///home/lennart/svn/public/pulseaudio/trunk@30 fefdeb5f-60dc-0310-8127-8f9354f1896f

+++ src/sample.h
@@ -19,7 +19,7 @@ enum sample_format {
 struct sample_spec {
     enum sample_format format;
     uint32_t rate;
-    uint32_t channels;
+    uint8_t channels;
 };
 
 #define DEFAULT_SAMPLE_SPEC default_sample_spec
